requirements.txt
accelerate==1.0.1
certifi==2024.8.30
charset-normalizer==3.4.0
controlnet-aux==0.0.9
diffusers==0.31.0
einops==0.8.0
filelock==3.16.1
fsspec==2024.10.0
huggingface-hub==0.26.1
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==11.0.0
psutil==6.1.0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.2.0
sympy==1.13.1
tifffile==2024.9.20
timm==0.6.7
tokenizers==0.20.1
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.5
transformers==4.45.2
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.2.3
zipp==3.20.2
